relocated data_process file
