+++ utils/tweets_data.json
@@ -1,0 +1,150 @@
+{
+    "_id":null,
+    "user_id":null,
+    "username":null,
+    "start_date":null,
+    "end_date":null,
+    "followings":0,
+    "followers":0,
+    "total_tweets":0,
+    "mentions":0,
+
+    "media_type":{
+        "text":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        },
+        "video":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        },
+        "photo":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        },
+        "link":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        },
+        "polls":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        },
+        "animated_gif":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        }
+    },
+    "highlights":{
+        "retweet_count":{"id":null,"count":0},
+        "reply_count":{"id":null,"count":0},
+        "like_count":{"id":null,"count":0},
+        "quote_count":{"id":null,"count":0},
+        "impression_count":{"id":null,"count":0}
+    },
+    "type": {
+        "retweeted":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0
+        },
+        "replied_to":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        },
+        "quoted":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        },
+        "original":{
+            "count" : 0,
+            "last" : 0,
+            "interval" : 0,
+            "metrics":{
+                "retweet_count":0,
+                "reply_count":0,
+                "like_count":0,
+                "quote_count":0,
+                "impression_count":0
+            }
+        }
+
+    },
+
+    "total" : {
+        "count" : 0,
+        "last" : 0,
+        "interval" : 0,
+        "metrics":{
+            "retweet_count":[],
+            "reply_count":[],
+            "like_count":[],
+            "quote_count":[],
+            "impression_count":[]
+        }
+    }
+}
